Found bad bug in calculation of C6. It is corrected using a new quantity in atomdata.f90.
The bug was that I was using the population of only the outest shell in order to calculate 
the C6 coefficients for aim. It was very bad since the coefficents was different alse if the charge
of the atom was 0.


Moreover I added an integer variable in precision.f90 named kch that is useful to define
character variables with always the same dimension.