First semplified version.
This version is "resoneably" bug free.

In this version the b parameter is not wheigted for each atom.
All the stuff to obtain more than one results at the time have been removed:
now the results correspond at the  A-D3 approach since it has demonstated to
be the best for profiles.

This is the version called dDMCv2 in some of my reports.